FIX: calc exp

+++ src/PLMData.sol
@@ -71,7 +71,7 @@ contract PLMData is IPLMData {
         pure
         returns (uint256)
     {
-        return charInfo.level**3;
+        return uint256(charInfo.level)**3;
     }
 
     function numOddsCharacterType() external view returns (uint256) {
